Spendenbescheinigung auch in Navigationsmenü
